+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
+- Grammar ANTLR for the calculator, supporting infix, prefix and postfix notations.
+- Parser and Visitor for the grammar, allowing to create an Expression tree from a string input.
 - Support for Unary operations
 - Trigonometric functions: sine, cosine, tangent and their inverses.
 - Others functions: logarithm, square root.